Sort modes and modalities
x @ mode1 mode2 and x @ mode2 mode1 are equivalent, as are things like `value
mod mode1 mode2` and `value mod mode2 mode1`. It'll be nice to have a canonical
order on these to avoid thrashing and inconsistency in code with modalities and
mode annotations. I decided to just go with sorting by String.compare for now,
rather than trying to impose any opinions on the order.

Signed-off-by: Aspen Smith <aspsmith@janestreet.com>